services/ai-ml/requirements.txt
fastapi>=0.110.0

uvicorn[standard]>=0.29.0
 
pydantic>=2.6.0

pydantic-settings>=2.2.1
 
structlog>=24.1.0

httpx>=0.27.0
 
google-genai>=0.3.0
 
numpy

redis>=5.0.0

tenacity>=8.2.3
 
google-generativeai

pdfplumber>=0.10.3

pypdf>=3.17.4

pdf2image>=1.17.0

pillow>=10.2.0

pytesseract>=0.3.10
 
python-docx>=1.1.0
 
sqlalchemy[asyncio]>=2.0.0

asyncpg>=0.29.0

psycopg2-binary>=2.9.9
 
pytest>=8.1.1

pytest-mock>=3.12.0
 